# split-bill-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ConfirmOCRRequest:
    properties:
      discount:
        type: number
      items:
        items:
          $ref: '#/definitions/models.ParsedItem'
        type: array
      paid_by:
        type: string
      service_fee:
        type: number
      split_among:
        description: user IDs for equal split
        items:
          type: string
        type: array
      split_type:
        type: string
      tax:
        type: number
      title:
        type: string
      total:
        type: number
    required:
    - items
    - paid_by
    - split_type
    - title
    - total
    type: object
...
  models.OCRResultResponse:
    properties:
      confidence_score:
        type: number
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      parsed_discount:
        type: number
      parsed_items:
        items:
          $ref: '#/definitions/models.ParsedItem'
        type: array
      parsed_service_fee:
        type: number
      parsed_tax:
        type: number
      parsed_total:
        type: number
      processing_time_ms:
        type: integer
      raw_text:
        type: string
      status:
        $ref: '#/definitions/models.OCRStatus'
    type: object
  models.OCRStatus:
    enum:
    - processing
    - completed
    - failed
    - confirmed
    type: string
    x-enum-varnames:
    - OCRStatusProcessing
    - OCRStatusCompleted
    - OCRStatusFailed
    - OCRStatusConfirmed
  models.ParsedItem:
    properties:
      confidence:
        description: 0.0 - 1.0
        type: number
      name:
        type: string
      quantity:
        type: integer
      total_price:
        type: number
      unit_price:
        type: number
    type: object
...
  models.ScanReceiptFromBase64Request:
    properties:
      file_name:
        type: string
      group_id:
        type: string
      image_base64:
        type: string
    required:
    - group_id
    - image_base64
    type: object
  models.ScanReceiptRequest:
    properties:
      group_id:
        type: string
      image_url:
        type: string
    required:
    - group_id
    - image_url
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@splitbill.app
    name: Split Bill Support
  description: API server for Split Bill application - manage groups, bills, transactions,
    OCR receipt scanning, and payments
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Split Bill API
  version: 4.0.0
paths:
...
  /ocr/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Confirms parsed OCR results, optionally edits items, and creates
        a bill from the scan
      parameters:
      - description: OCR Result ID
        in: path
        name: id
        required: true
        type: string
      - description: Confirmed items and bill details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConfirmOCRRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Confirm OCR scan and create bill
      tags:
      - OCR
  /ocr/{id}/result:
    get:
      description: Retrieves a previously scanned OCR result by ID
      parameters:
      - description: OCR Result ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OCRResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get OCR result
      tags:
      - OCR
  /ocr/pending:
    get:
      description: Returns all pending (unconfirmed) OCR scans for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.OCRResultResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get pending OCR scans
      tags:
      - OCR
  /ocr/scan:
    post:
      consumes:
      - application/json
      description: Scans a receipt image from a URL using OCR and extracts items
      parameters:
      - description: Image URL and group ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ScanReceiptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OCRResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Scan receipt from URL
      tags:
      - OCR
  /ocr/scan-base64:
    post:
      consumes:
      - application/json
      description: Scans a receipt from a base64-encoded image using OCR and extracts
        items
      parameters:
      - description: Base64 image and group ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ScanReceiptFromBase64Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.OCRResultResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Scan receipt from base64 image
      tags:
      - OCR
...
  /upload/image:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image file (JPEG, PNG, WebP) with max 10MB size limit
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload image via multipart form
      tags:
      - Upload
  /upload/image-base64:
    post:
      consumes:
      - application/json
      description: Uploads a base64-encoded image (supports data URI format) with
        max 10MB size limit
      parameters:
      - description: Base64 image data
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Upload image via base64
      tags:
      - Upload
...
securityDefinitions:
  BearerAuth:
    description: 'Firebase Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"